+++ utils.go
@@ -1,8 +1,10 @@
 package timeago
 
 import (
+	"embed"
 	"encoding/json"
 	"errors"
+	"fmt"
 	"log"
 	"os"
 	"time"
@@ -12,17 +14,19 @@ func unixToTime(userDate int) time.Time {
 	return time.Unix(int64(userDate), 0)
 }
 
-func parseLangSet(fileName string) *LangSet {
-	content := readFile(fileName)
-
-	var res LangSet
-
-	err := json.Unmarshal(content, &res)
+func parseLangSet(filePath string, fs embed.FS) (*LangSet, error) {
+	content, err := langsFS.ReadFile(filePath)
 	if err != nil {
-		log.Fatalln(err)
+		return nil, fmt.Errorf("could not read language file: %w", err)
 	}
 
-	return &res
+	langSet := &LangSet{}
+
+	if err := json.Unmarshal(content, &langSet); err != nil {
+		return nil, fmt.Errorf("could not unmarshal language file: %w", err)
+	}
+
+	return langSet, nil
 }
 
 func readFile(filePath string) []byte {