+++ utils_test.go
@@ -7,7 +7,11 @@ import (
 
 func TestParseJsonIntoLang(t *testing.T) {
 	t.Run("Function returns Lang model with needed values", func(test *testing.T) {
-		res := parseLangSet("langs/ru.json")
+		res, err := parseLangSet("langs/ru.json", langsFS)
+
+		if err != nil {
+			t.Errorf("Function parseLangSet must return Lang model, but returned error %v", err)
+		}
 
 		if res.Ago != "назад" {
 			t.Errorf("Function needs to return model with value назад, but returned %v", res.Ago)