+++ utils.go
@@ -24,3 +24,18 @@ func parseLangSet(filePath string) (*LangSet, error) {
 
 	return langSet, nil
 }
+
+func isUnsignedInteger(s string) bool {
+	if len(s) == 0 {
+		return false
+	}
+
+	for i := 0; i < len(s); i++ {
+		c := s[i]
+		if c < '0' || c > '9' {
+			return false
+		}
+	}
+
+	return true
+}