+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release Notes v3
 
 ## v3.2.2 (2025-06-04)
+- Change so that all JSON files are included in the final binary by using `go:embed` functionality
 - Remove support for go version 1.13, 1.14, 1.15 because they do not have `go:embed`
 
 ## v3.2.1 (2024-12-06)